+++ src/inspect_ai/tool/__init__.py
@@ -19,10 +19,10 @@ from inspect_ai._util.deprecation import relocated_module_attribute
 from ._mcp import (
     MCPServer,
     mcp_connection,
+    mcp_server_http,
     mcp_server_sandbox,
     mcp_server_sse,
     mcp_server_stdio,
-    mcp_server_streamablehttp,
     mcp_tools,
 )
 from ._tool import Tool, ToolError, ToolResult, ToolSource, tool
@@ -67,7 +67,7 @@ __all__ = [
     "mcp_connection",
     "mcp_server_stdio",
     "mcp_server_sse",
-    "mcp_server_streamablehttp",
+    "mcp_server_http",
     "mcp_server_sandbox",
     "MCPServer",
     "Content",